refactor(web): admin service단 코드위치 변경, api 코드 이름변경 위치변경
